Tot funcionant, crea partides,sales y zones. Faltaria part logica

+++ risk_project_daw/src/main/java/org/project/intermodular/risk_project_daw/model/db/SalasDb.java
@@ -18,8 +18,8 @@ public class SalasDb {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long  id_sala;
-    private Integer jugador1;
-    private Integer jugador2;
-    private Integer jugador3;
-    private Integer jugador4;
+    private Long jugador1;
+    private Long jugador2;
+    private Long jugador3;
+    private Long jugador4;
 }
