+++ risk_project_daw/src/main/java/org/project/intermodular/risk_project_daw/model/dto/SalasInfo.java
@@ -12,10 +12,10 @@ import lombok.NoArgsConstructor;
 public class SalasInfo {
     
     private Long id_sala;
-    private Integer jugador1;
-    private Integer jugador2;
-    private Integer jugador3;
-    private Integer jugador4;
+    private Long jugador1;
+    private Long jugador2;
+    private Long jugador3;
+    private Long jugador4;
    
 
 }
